Inaccurate camera rotation fix
Hello, I've fixed this issue:
https://github.com/codewithtom/godot-fps/issues/1
Along with some changes to make the controller a little bit better.
Hopefully this controller will be the basis for new Godoters who just wants to start up their journey without having to code everything from scratch. or just for prototyping.

Features:
+ The camera rotation is more accurate and smoother, since the movement is handled in the process function instead of the input function.
+ The camera rotation is limited by a vertical FOV instead of hard coded values.
EX: (-90°; 90°) = 180° vertical FOV
+ Changed the camera vertical FOV to 140° because it's more common in FPS games.
+ Added a way to re-capture the mouse by left clicking the window or by pressing the "ESC" key again.
+ Minor adjustments.

Potential issues:
As long as am concerned, the only thing that doesn't feel right with in the current implementation is that when you re-capture the mouse by clicking the window. The weapon fires too.
I might fix that too if this changes got accepted.
Thank you for the awesome tutorial by the way!